requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
jinja2==3.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.7
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pandas==2.0.0
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
groq==0.4.1
python-jose[cryptography]==3.3.0
markdown2==2.4.8 